Update traning_example.rs - i64->f64 in x_arr

+++ src/data_preprocess.rs
@@ -3,15 +3,15 @@ use rand::Rng;
 // use std::error::Error;
 
 pub fn split_data_randomly(
-    x: Vec<i64>, 
+    x: Vec<f64>, 
     y: Vec<f64>, 
     min_chunk_size: usize, 
     max_chunk_size: usize
-) -> Vec<(Vec<i64>, Vec<f64>)> {
+) -> Vec<(Vec<f64>, Vec<f64>)> {
     let mut rng = rand::thread_rng();
     
     // Combine x and y into a tuple vector and shuffle
-    let mut data: Vec<(i64, f64)> = x.into_iter().zip(y.into_iter()).collect();
+    let mut data: Vec<(f64, f64)> = x.into_iter().zip(y.into_iter()).collect();
     data.shuffle(&mut rng);
     
     let mut chunks = Vec::new();
@@ -23,10 +23,10 @@ pub fn split_data_randomly(
             .min(remaining_data.len());
 
         // Split off a chunk of the data
-        let chunk: Vec<(i64, f64)> = remaining_data.drain(0..chunk_size).collect();
+        let chunk: Vec<(f64, f64)> = remaining_data.drain(0..chunk_size).collect();
 
         // Separate the chunk into x and y again
-        let (chunk_x, chunk_y): (Vec<i64>, Vec<f64>) = chunk.into_iter().unzip();
+        let (chunk_x, chunk_y): (Vec<f64>, Vec<f64>) = chunk.into_iter().unzip();
         chunks.push((chunk_x, chunk_y));
     }
