+++ src/traning_example.rs
@@ -1,10 +1,8 @@
 use csv::ReaderBuilder;
 use std::error::Error;
-// use linfa::prelude::*;
-// use linfa_linear::LinearRegression;
-// use rand::seq::SliceRandom;
 
-pub fn read_csv(file_path: &str) -> Result<(Vec<i64>, Vec<f64>), Box<dyn Error>> {
+
+pub fn read_csv(file_path: &str) -> Result<(Vec<f64>, Vec<f64>), Box<dyn Error>> {
     let mut reader = ReaderBuilder::new()
     .has_headers(true)
     .from_path(file_path)?;
@@ -16,7 +14,7 @@ for (i, result) in reader.records().enumerate() {
     match result {
         Ok(record) => {
             println!("Record {}: {:?}", i + 1, record);
-            let x: i64 = match record[0].parse() {
+            let x: f64 = match record[0].parse() {
                 Ok(val) => val,
                 Err(e) => {
                     eprintln!("Error parsing 'x' at record {}: {}", i + 1, e);
@@ -44,22 +42,3 @@ for (i, result) in reader.records().enumerate() {
 
 Ok((x_values, y_values))
 }
-
-// fn train_model_in_chunks(x: Vec<f64>, y: Vec<f64>, chunk_size: usize) -> LinearRegression<f64> {
-//     let mut rng = rand::thread_rng();
-//     let mut x_chunked = x.chunks(chunk_size).collect::<Vec<_>>();
-//     let mut y_chunked = y.chunks(chunk_size).collect::<Vec<_>>();
-
-//     // Shuffle chunks
-//     x_chunked.shuffle(&mut rng);
-//     y_chunked.shuffle(&mut rng);
-
-//     let mut model = LinearRegression::default();
-
-//     for (x_chunk, y_chunk) in x_chunked.iter().zip(y_chunked.iter()) {
-//         let dataset = linfa::Dataset::from((x_chunk.to_vec(), y_chunk.to_vec()));
-//         model.fit(&dataset).unwrap();
-//     }
-
-//     model
-// }
